:sparkles:  基本实现了jmm与nmm的ipc通讯

+++ desktop/scripts/sync-to-android-assets.cjs
@@ -8,7 +8,11 @@ const android_assets_dir = path.join(
 );
 const origin_dir = path.join(__dirname, "../");
 
-const tasks = ["bundle"].map((dirname) => {// , "cot"
+const tasks = [
+  //
+  "bundle",
+  "cot",
+].map((dirname) => {
   return {
     from: path.join(origin_dir, dirname),
     to: path.join(android_assets_dir, dirname),
@@ -20,7 +24,9 @@ const doSync = () => {
     if (fs.existsSync(task.to)) {
       fs.rmSync(task.to, { recursive: true });
     }
-    fs.cpSync(task.from, task.to, { recursive: true });
+    if (fs.existsSync(task.from)) {
+      fs.cpSync(task.from, task.to, { recursive: true });
+    }
   }
   console.log("synced");
 };
@@ -30,6 +36,7 @@ if (process.argv.includes("--watch")) {
   const debounceSync = debounce(doSync, 500);
   for (const task of tasks) {
     fs.watch(task.from, { recursive: true }, (event) => {
+      console.log(event)
       debounceSync();
     });
   }
